debug

+++ include/linux_parser.h
@@ -20,7 +20,7 @@ const std::string kPasswordPath{"/etc/passwd"};
 
 // System
 float MemoryUtilization();
-long UpTime();
+long long int UpTime();
 std::vector<int> Pids();
 int TotalProcesses();
 int RunningProcesses();
@@ -41,17 +41,17 @@ enum CPUStates {
   kGuestNice_
 };
 std::vector<int> CpuUtilization();
-long Jiffies();
-long ActiveJiffies();
-long ActiveJiffies(int pid);
-long IdleJiffies();
+long long int Jiffies();
+long long int ActiveJiffies();
+long long int ActiveJiffies(int pid);
+long long int IdleJiffies();
 
 // Processes
 std::string Command(int pid);
 std::string Ram(int pid);
 std::string Uid(int pid);
 std::string User(int pid);
-long int UpTime(int pid);
+long long int UpTime(int pid);
 };  // namespace LinuxParser
 
 #endif
